requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
kucoin-futures-python>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
sqlite3>=3.35.0
matplotlib>=3.4.0
ta>=0.7.0